refactor: rename package name
