+++ app/src/main/kotlin/com/wire/android/ui/calling/controlbuttons/AcceptButton.kt
@@ -1,4 +1,4 @@
-package com.wire.android.ui.calling.controlButtons
+package com.wire.android.ui.calling.controlbuttons
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.height
@@ -8,12 +8,11 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import com.wire.android.R
 import com.wire.android.ui.common.dimensions
 
 @Composable
-fun DeclineButton(buttonClicked: () -> Unit) {
+fun AcceptButton(buttonClicked: () -> Unit) {
     IconButton(
         modifier = Modifier
             .height(dimensions().initiatingCallHangUpButtonSize)
@@ -22,15 +21,9 @@ fun DeclineButton(buttonClicked: () -> Unit) {
     ) {
         Image(
             painter = painterResource(
-                id = R.drawable.ic_decline
+                id = R.drawable.ic_accept
             ),
-            contentDescription = stringResource(id = R.string.content_description_calling_decline_call)
+            contentDescription = stringResource(id = R.string.content_description_calling_accept_call)
         )
     }
-}
-
-@Preview
-@Composable
-fun ComposableDeclineButtonPreview() {
-    DeclineButton(buttonClicked = { })
 }
