+++ app/src/main/kotlin/com/wire/android/ui/calling/controlbuttons/CallOptionsControls.kt
@@ -1,4 +1,4 @@
-package com.wire.android.ui.calling.controlButtons
+package com.wire.android.ui.calling.controlbuttons
 
 import android.widget.Toast
 import androidx.compose.foundation.layout.Arrangement
