+++ app/src/main/kotlin/com/wire/android/ui/calling/controlbuttons/CameraFlipButton.kt
@@ -1,4 +1,4 @@
-package com.wire.android.ui.calling.controlButtons
+package com.wire.android.ui.calling.controlbuttons
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.height
@@ -6,6 +6,10 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
@@ -14,34 +18,34 @@ import com.wire.android.R
 import com.wire.android.ui.theme.wireDimensions
 
 @Composable
-fun MicrophoneButton(
-    isMuted: Boolean,
-    onMicrophoneButtonClicked: () -> Unit
+fun CameraFlipButton(
+    isCameraFlipped: Boolean = false,
+    onCameraFlipButtonClicked: () -> Unit
 ) {
+    var isCameraFlipped by remember { mutableStateOf(isCameraFlipped) }
+
     IconButton(
         modifier = Modifier
             .width(MaterialTheme.wireDimensions.defaultCallingControlsSize)
             .height(MaterialTheme.wireDimensions.defaultCallingControlsSize),
-        onClick = onMicrophoneButtonClicked
+        onClick = onCameraFlipButtonClicked
     ) {
         Image(
             painter = painterResource(
-                id = if (isMuted) {
-                    R.drawable.ic_muted
-                } else {
-                    R.drawable.ic_unmuted
-                }
+                id = if (isCameraFlipped)
+                    R.drawable.ic_flip_camera_on
+                else R.drawable.ic_flip_camera_off
             ),
             contentDescription = stringResource(
-                id = if (isMuted) R.string.content_description_calling_unmute_call
-                else R.string.content_description_calling_mute_call
-            )
+                id = if (isCameraFlipped) R.string.content_description_calling_flip_camera_on
+                else R.string.content_description_calling_flip_camera_off
+            ),
         )
     }
 }
 
 @Preview
 @Composable
-fun ComposableMicrophoneButtonPreview() {
-    MicrophoneButton(isMuted = true, onMicrophoneButtonClicked = { })
+fun ComposableCameraFlipButtonPreview() {
+    CameraFlipButton(onCameraFlipButtonClicked = { })
 }
