+++ app/src/main/kotlin/com/wire/android/ui/calling/controlbuttons/HangUpButton.kt
@@ -1,4 +1,4 @@
-package com.wire.android.ui.calling.controlButtons
+package com.wire.android.ui.calling.controlbuttons
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.height
