+++ app/src/main/kotlin/com/wire/android/ui/calling/controlbuttons/MicrophoneButton.kt
@@ -1,4 +1,4 @@
-package com.wire.android.ui.calling.controlButtons
+package com.wire.android.ui.calling.controlbuttons
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.height
@@ -6,10 +6,6 @@ import androidx.compose.foundation.layout.width
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
@@ -18,34 +14,34 @@ import com.wire.android.R
 import com.wire.android.ui.theme.wireDimensions
 
 @Composable
-fun CameraFlipButton(
-    isCameraFlipped: Boolean = false,
-    onCameraFlipButtonClicked: () -> Unit
+fun MicrophoneButton(
+    isMuted: Boolean,
+    onMicrophoneButtonClicked: () -> Unit
 ) {
-    var isCameraFlipped by remember { mutableStateOf(isCameraFlipped) }
-
     IconButton(
         modifier = Modifier
             .width(MaterialTheme.wireDimensions.defaultCallingControlsSize)
             .height(MaterialTheme.wireDimensions.defaultCallingControlsSize),
-        onClick = onCameraFlipButtonClicked
+        onClick = onMicrophoneButtonClicked
     ) {
         Image(
             painter = painterResource(
-                id = if (isCameraFlipped)
-                    R.drawable.ic_flip_camera_on
-                else R.drawable.ic_flip_camera_off
+                id = if (isMuted) {
+                    R.drawable.ic_muted
+                } else {
+                    R.drawable.ic_unmuted
+                }
             ),
             contentDescription = stringResource(
-                id = if (isCameraFlipped) R.string.content_description_calling_flip_camera_on
-                else R.string.content_description_calling_flip_camera_off
-            ),
+                id = if (isMuted) R.string.content_description_calling_unmute_call
+                else R.string.content_description_calling_mute_call
+            )
         )
     }
 }
 
 @Preview
 @Composable
-fun ComposableCameraFlipButtonPreview() {
-    CameraFlipButton(onCameraFlipButtonClicked = { })
+fun ComposableMicrophoneButtonPreview() {
+    MicrophoneButton(isMuted = true, onMicrophoneButtonClicked = { })
 }
