+++ app/src/main/kotlin/com/wire/android/ui/calling/ongoing/OngoingCallScreen.kt
@@ -31,11 +31,11 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import com.wire.android.R
 import com.wire.android.ui.calling.ConversationName
 import com.wire.android.ui.calling.SharedCallingViewModel
-import com.wire.android.ui.calling.controlButtons.CameraButton
-import com.wire.android.ui.calling.controlButtons.CameraFlipButton
-import com.wire.android.ui.calling.controlButtons.HangUpButton
-import com.wire.android.ui.calling.controlButtons.MicrophoneButton
-import com.wire.android.ui.calling.controlButtons.SpeakerButton
+import com.wire.android.ui.calling.controlbuttons.CameraButton
+import com.wire.android.ui.calling.controlbuttons.CameraFlipButton
+import com.wire.android.ui.calling.controlbuttons.HangUpButton
+import com.wire.android.ui.calling.controlbuttons.MicrophoneButton
+import com.wire.android.ui.calling.controlbuttons.SpeakerButton
 import com.wire.android.ui.calling.model.UICallParticipant
 import com.wire.android.ui.calling.ongoing.participantsview.VerticalCallingPager
 import com.wire.android.ui.common.SecurityClassificationBanner
